apollo/requirements.txt
fastapi
uvicorn
requests
argparse
python-dotenv
pysqlite3